backend-python/src/requirements.txt
fastapi
uvicorn
python-dotenv
sqlmodel
pymysql   # or mysql-connector-python — but pymysql is faster/common with SQLAlchemy/SQLModel
jose
argon2-cffi
